Add files via upload
I made this in VS, but I don't know if it works or not. I followed instructions on how to make this, and cross referenced to see that they were the exact same (apart from changing the size of some of the shapes.) My live view extension wasn't giving me a live view, so I don't know if it was frozen or just doesn't function. The eyes are supposed to follow the mouse.